api-v2/services/common/admin-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.2
jose==3.3.0

